changement des float en double + test main

+++ src/main/java/insa/aubin/devisbatiment/Fenetre.java
@@ -1,7 +1,7 @@
 package insa.aubin.devisbatiment;
 
 public class Fenetre extends Ouverture {
-    private static final float COTE_FENETRE = 1.20f;  //Par convention, les nomes des constantes s'écrivent comme : UPPER_SNAKE_CASE
+    private static final double COTE_FENETRE = 1.20f;  //Par convention, les nomes des constantes s'écrivent comme : UPPER_SNAKE_CASE
 
     public Fenetre(String idFenetre){
         super(idFenetre, COTE_FENETRE, COTE_FENETRE);
